Tablas y código ajustado
