+++ src/modules/grupos/application/dtos/create-group.dto.ts
@@ -1,0 +1,13 @@
+import { IsNotEmpty, IsOptional, IsString, MaxLength } from 'class-validator';
+
+export class CrearGrupoDTO {
+  @IsNotEmpty({ message: 'El nombre es obligatorio' })
+  @IsString()
+  @MaxLength(64)
+  nombre: string;
+
+  @IsOptional()
+  @IsString()
+  @MaxLength(255)
+  descripcion?: string;
+}
