+++ src/modules/grupos/application/dtos/update-group.dto.ts
@@ -1,0 +1,13 @@
+import { IsOptional, IsString, MaxLength } from 'class-validator';
+
+export class UpdateGroupDTO {
+  @IsOptional()
+  @IsString()
+  @MaxLength(64)
+  nombre?: string;
+
+  @IsOptional()
+  @IsString()
+  @MaxLength(255)
+  descripcion?: string;
+}
