+++ src/modules/grupos/domain/interfaces/grupo-repository.interface.ts
@@ -1,0 +1,9 @@
+import { Suscripcion } from "../entities/subscription";
+
+export abstract class IGrupoRepository {
+  abstract buscarPorId(id: string): Promise<Suscripcion | null>;
+  abstract crear(suscripcion: Suscripcion): Promise<Suscripcion>;
+  abstract actualizar(id: string, suscripcion: Suscripcion): Promise<Suscripcion>;
+  abstract eliminar(id: string): Promise<void>;
+  abstract obtenerTodas(): Promise<Suscripcion[]>;
+}
