--- src/modules/suscripciones/application/dtos/create-subscription.dto.ts
@@ -1,13 +1,0 @@
-import { IsNotEmpty, IsOptional, IsString, MaxLength } from 'class-validator';
-
-export class CrearSuscripcionDTO {
-  @IsNotEmpty({ message: 'El nombre es obligatorio' })
-  @IsString()
-  @MaxLength(64)
-  nombre: string;
-
-  @IsOptional()
-  @IsString()
-  @MaxLength(255)
-  descripcion?: string;
-}
