--- src/modules/usuario-suscripcion/domain/entities/usuarioSuscripcion.ts
@@ -1,23 +1,0 @@
-export class UsuarioSuscripcion { 
-    readonly id: string;
-    readonly usuarioId: string;
-    readonly suscripcionId: string;
-
-    constructor(
-        id:string,
-        usuarioId: string,
-        suscripcionId: string
-    ){
-        this.id = id;
-        this.usuarioId = usuarioId;
-        this.suscripcionId = suscripcionId;
-    }
-
-    value(){
-        return {
-            id: this.id,
-            usuarioId:this.usuarioId,
-            suscripcionId : this.suscripcionId
-        }
-    }
-}
