--- src/modules/usuario-suscripcion/domain/interfaces/usuario-suscripcion-repository.interface.ts
@@ -1,10 +1,0 @@
-import { UsuarioSuscripcion } from "../entities/usuarioSuscripcion";
-
-export abstract class IUsuarioSuscripcionRepository {
-  abstract crear(usuarioSuscripcion: UsuarioSuscripcion): Promise<UsuarioSuscripcion>;
-
-  abstract existeRelacion(usuarioId: string, suscripcionId: string): Promise<boolean>;
-  //Agregando la Eliminacion
-  abstract eliminar(usuarioId: string, suscripcionId: string): Promise<void>;
-
-}
