--- src/modules/usuario-suscripcion/usuario-suscripcion.module.ts
@@ -1,21 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PrismaService } from 'src/core/databases/prisma.service';
-import { UsuarioSuscripcionPrismaRepository } from './infraestructure/repositories/usuario-suscripcion-prisma.repository';
-import { IUsuarioSuscripcionRepository } from './domain/interfaces/usuario-suscripcion-repository.interface';
-import { AsociarUsuarioUseCase } from './application/use-cases/asociar-usuario.use-case';
-import { UsuarioSuscripcionController } from './infraestructure/controllers/usuario-suscripcion.controller';
-import { EliminarMiembroUseCase } from './application/use-cases/eliminarMiembro-use.case';
-
-@Module({
-  controllers: [UsuarioSuscripcionController],
-  providers: [
-    PrismaService,
-    AsociarUsuarioUseCase,
-    EliminarMiembroUseCase,
-    {
-      provide: IUsuarioSuscripcionRepository,
-      useClass: UsuarioSuscripcionPrismaRepository,
-    },
-  ],
-})
-export class UsuarioSuscripcionModule {}
